% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfunite.R
\name{pdfunite}
\alias{pdfunite}
\title{Interact with command line utility pdfunite to combine pdfs into a single file.}
\usage{
pdfunite(input, output, check = TRUE, verbose = TRUE)
}
\arguments{
\item{input}{Character vector of separate pdf files to join.}

\item{output}{Character string representing output pdf file name.}

\item{check}{Logical, whether or not to invoke Sys.which to check for existance of pdfunite command line utility.}

\item{verbose}{Logical, whether or not to print command before running.}
}
\description{
Interact with command line utility pdfunite to combine pdfs into a single file.
}
