% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfuniter.R
\docType{package}
\name{pdfuniter}
\alias{pdfuniter}
\alias{pdfuniter-package}
\title{pdfuniter: R package to interact with pdfunite command line utility.}
\description{
pdfuniter: R package to interact with pdfunite command line utility.
}
