% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfunite.R
\name{pdfunite}
\alias{pdfunite}
\title{Interact with command line utility pdfunite to combine pdfs into a single file.}
\usage{
pdfunite(input, output, check = TRUE, verbose = TRUE)
}
\arguments{
\item{input}{Character vector of separate pdf files to join.}

\item{output}{Character string representing output pdf file name.}

\item{check}{Logical, whether or not to invoke Sys.which to check for existance of pdfunite command line utility.}

\item{verbose}{Logical, whether or not to print command before running.}
}
\description{
Interact with command line utility pdfunite to combine pdfs into a single file.
}
\examples{
pdf("page1.pdf")
plot(x = c(1:10), y = rnorm(10))
dev.off()
pdf("page2.pdf")
plot(x = c(1:10), y = c(1:10))
dev.off()
pdfunite(c("page1.pdf", "page2.pdf"), "full_document.pdf")

}
